bugfix: the maximum possible fitness of an application without priority should be "max reachable RTT"
The maximum possible fitness value contributed by every application should be the same, which is the “max reachable RTT” in my design.
However, in the old code, every application depending on others has 3 parts of fitness, i.e., network part, this application part, and dependent application part, and every part has the maximum possible fitness “max reachable RTT”, so the sum is 3 * “max reachable RTT”.
Therefore, in the old code,
- if an application depends on others, its maximum possible fitness is 3 * “max reachable RTT”; (MRRTT * 3 * depNum / depNum = MRRTT * 3)
- if it does not depend on others, its maximum possible fitness is “max reachable RTT”. (MRRTT)
Hence, we should use 3 to divide the fitnesses of applications depending on others, so that all applications have the same weight if their priorities are the same.